requirements.txt
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
discord-webhook>=1.3.0
sofascore-wrapper>=1.1.0
playwright>=1.42.0
schedule>=1.2.0
aiohttp>=3.9.0